Rename FramebufferCommon.cpp/h to FramebufferManagerCommon.cpp/h for consistency
