+++ GPU/Directx9/FramebufferDX9.h
@@ -26,7 +26,7 @@
 // pixel data.
 
 #include "GPU/GPUCommon.h"
-#include "GPU/Common/FramebufferCommon.h"
+#include "GPU/Common/FramebufferManagerCommon.h"
 
 namespace DX9 {
