+++ GPU/Vulkan/FramebufferVulkan.cpp
@@ -41,7 +41,7 @@
 #include "GPU/GPUState.h"
 
 #include "GPU/Common/TextureDecoder.h"
-#include "GPU/Common/FramebufferCommon.h"
+#include "GPU/Common/FramebufferManagerCommon.h"
 #include "GPU/Debugger/Stepping.h"
 #include "GPU/Vulkan/FramebufferVulkan.h"
 #include "GPU/Vulkan/DrawEngineVulkan.h"
