+++ UI/DisplayLayoutScreen.cpp
@@ -31,7 +31,7 @@
 #include "Core/ConfigValues.h"
 #include "Core/System.h"
 #include "DisplayLayoutEditor.h"
-#include "GPU/Common/FramebufferCommon.h"
+#include "GPU/Common/FramebufferManagerCommon.h"
 
 static const int leftColumnWidth = 200;
 static const float orgRatio = 1.764706f;
